#  [WIP] Attempt splitting the reduction part