Attempt adding a SplitOp to the reduction part to allow `vector<4>` more generally even when the reduction size is not a multiple of `4`.

Unfortunately this results in increased shared memory usage which tanks the performance.
This is an illustration of the need for a post-vector distribution privatization pass to remove the buffers that have been inserted for private use.
    
Repro:
```
(benchmark-transform-create -r  ../iree-samples/transform_dialect/benchmark_linalg_reductions.stub.mlir   reduction_2d_elementwise_static f32 1000 3232)
```
    
Before:
```
467.30ms  52.387us           (1000 1 1)       (512 1 1)        30        0B  2.0664KB  NVIDIA GeForce          1        13                     -                -  reduction_2d_elementwise_static_dispatch_0_generic_1000x3232 [29]
With transform dialect: reduction_2d_elementwise_static --function_input="1000x3232xf32=1" P50: 52035.000 ns 69.00647641010858076294 GElements/s
```
    
After:
```
433.36ms  115.59us           (1000 1 1)       (512 1 1)        30        0B  4.0664KB  NVIDIA GeForce          1        13                     -                -  reduction_2d_elementwise_static_dispatch_0_generic_1000x3232 [29]
With transform dialect: reduction_2d_elementwise_static --function_input="1000x3232xf32=1" P50: 115940.00 ns 30.97077798861480075901 GElements/s
```